pyproject.toml
[project]
name = "zoombuild"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
	"python-dotenv>=1.01",
	"tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.4",
    "pytest",
]


[project.scripts]
zip = "copier.py"

[tool.ruff]
line-length = 99

[tool.ruff.lint]
# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

